APIs_and_Microservices/04_Securing/requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
apispec==6.3.0
black==23.10.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
config==0.5.1
cryptography==41.0.5
fastapi==0.104.0
Flask==2.3.3
flask-smorest==0.42.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jq==1.6.0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mypy-extensions==1.0.0
nodejs==0.1.1
optional-django==0.1.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
webargs==8.3.0
Werkzeug==2.3.7
